Use double instead of float.

+++ fftme/mfft1.cl
@@ -6,9 +6,11 @@ unsigned int uintlog2(unsigned int n)
   return r;
 }
 
-void swap(__global float *f, const unsigned int a, const unsigned int b)
+//#define double double;
+
+void swap(__global double *f, const unsigned int a, const unsigned int b)
 {
-  float temp[2]={f[2*a],f[2*a+1]};
+  double temp[2]={f[2*a],f[2*a+1]};
   f[2*a]   = f[2*b];
   f[2*a+1] = f[2*b+1];
   f[2*b]   = temp[0];
@@ -37,7 +39,7 @@ unsigned int bitreverse(const unsigned int k, const unsigned int log2ny)
   return kr;
 }
 
-void unshuffle( __global float *fx, const unsigned int ny)
+void unshuffle( __global double *fx, const unsigned int ny)
 {
   const unsigned int log2ny = uintlog2(ny);
   for(unsigned int k = 0; k < ny; ++k) {
@@ -73,7 +75,7 @@ unsigned int even(const unsigned int l2n,
 }
 
 __kernel void mfft1(unsigned int nx, unsigned int mx, 
-		    unsigned int ny, __global float *f)
+		    unsigned int ny, __global double *f)
 {
   /* const unsigned int l2n=log2(n); */
 
@@ -81,7 +83,7 @@ __kernel void mfft1(unsigned int nx, unsigned int mx,
 
   const unsigned int log2ny = uintlog2(ny);
   
-  const float PI=4.0*atan(1.0);
+  const double PI=4.0*atan(1.0);
   unsigned int kb[32]; // this is too big, but it compiles!
   /* unsigned int *kb=new unsigned int[log2ny]; */
   const unsigned int kymax = ny / 2;
@@ -91,7 +93,7 @@ __kernel void mfft1(unsigned int nx, unsigned int mx,
   const unsigned int ixstart = mx * idx;
   const unsigned int ixstop = min(ixstart + mx, nx);
   for(unsigned int ix = ixstart; ix < ixstop; ++ix) {
-    float *fx = f + 2 * (ix * ny);
+    double *fx = f + 2 * (ix * ny);
   
     for(unsigned int iy = 0; iy < log2ny; ++iy) {
       for(unsigned int ky = 0; ky < kymax; ++ky) {
@@ -100,18 +102,18 @@ __kernel void mfft1(unsigned int nx, unsigned int mx,
 	const unsigned int ke = even(log2ny, iy, kb);
 	const unsigned int ko = ke + twojy;
 
-	float fe[2] = {fx[2*ke], fx[2*ke+1]};
-	float fo[2] = {fx[2*ko], fx[2*ko+1]};
+	double fe[2] = {fx[2*ke], fx[2*ke+1]};
+	double fo[2] = {fx[2*ko], fx[2*ko+1]};
       
 	// TODO: move w to a lookup table (in local memory?)
-	/* const float arg = -2.0 * PI * ky * iy / (float)ny; */
-	const float arg = -2.0 * PI * ke / (2.0 * twojy);
-	const float w[2] = {cos(arg), sin(arg)};
+	/* const double arg = -2.0 * PI * ky * iy / (double)ny; */
+	const double arg = -2.0 * PI * ke / (2.0 * twojy);
+	const double w[2] = {cos(arg), sin(arg)};
       
 	fx[2*ke]   = fe[0] + fo[0];
 	fx[2*ke+1] = fe[1] + fo[1];
 
-	float t[2] = {fe[0] - fo[0], fe[1] - fo[1]};
+	double t[2] = {fe[0] - fo[0], fe[1] - fo[1]};
       
 	fx[2*ko]   = w[0]*t[0] - w[1]*t[1];
 	fx[2*ko+1] = w[1]*t[0] + w[0]*t[1];
